Removes loadbalancers for OC 4.3 tests to eliminate hangs
OpenShift Version 4.3 has a known bug whereby deletion of services
that use LoadBalancers (or namespaces that contain such services)
can cause indefinite hanging. Apparently there are finalizers used
on these services to force cleanup of loadbalancer resources, and
the finalizers are never deleted from the service:
  https://bugzilla.redhat.com/show_bug.cgi?id=1798282

This change disables the use of LoadBalancers for the OpenShift 4.3
testing as a temporary workaround.